requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
azure-ai-contentsafety==1.0.0
azure-core==1.38.0
python-dotenv==1.2.1
requests==2.32.5
pypdf==6.6.2
python-docx==1.2.0